chore: remove unnecessary typcast

+++ src/utils.c
@@ -36,7 +36,7 @@ int get_num_digits(__off_t n) {
 char *double_if_of(char *buf, size_t idx, size_t addition, size_t *size) {
 	char *tmp = NULL;
 
-	if (idx + (uint)addition > *size) {
+	if (idx + addition > *size) {
 		*size *= 2;
 
 		if ((tmp = realloc(buf, *size)) == NULL) {
